# server/openapi.yml
openapi: 3.0.3
info:
  title: Venue Booking System - Identity & Auth API
  description: Production OpenAPI specification for US-B1 (Auth & Session) and US-B2 (Auto-registration & Role Elevation).
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local Development Server

paths:
  /auth/login:
    post:
      summary: Institutional SSO Login / Auto-registration
      description: Authenticates a user via institutional email (@iiml.ac.in). Auto-registers first-time users as 'Booker'.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: student@iiml.ac.in
                name:
                  type: string
                  example: Jane Doe
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          description: Invalid parameters supplied
        '403':
          description: Access denied (non-institutional email domain)
          content:
            application/json:
              example:
                error: Access denied. Only @iiml.ac.in email addresses are permitted.
        '500':
          description: Server error

  /auth/me:
    get:
      summary: Get Current Authenticated User
      description: Returns the user object and current role resolved from the database.
      operationId: getCurrentUser
      security:
        - SessionTokenAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Successfully fetched current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Missing or invalid session token
        '403':
          description: Forbidden (domain check failed)

  /auth/roles/elevate:
    put:
      summary: Elevate User Role (Admin Only)
      description: Updates a user's role and writes an immutable audit log entry in the same atomic transaction.
      operationId: elevateUserRole
      security:
        - SessionTokenAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - targetUserId
                - newRole
              properties:
                targetUserId:
                  type: string
                  format: uuid
                  example: 123e4567-e89b-12d3-a456-426614174000
                newRole:
                  type: string
                  enum: [Booker, Approver, Admin]
                  example: Approver
      responses:
        '200':
          description: Role elevated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User role updated successfully.
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid request body or unsupported role
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden - Action requires Admin privilege
        '404':
          description: Target user not found

components:
  securitySchemes:
    SessionTokenAuth:
      type: apiKey
      in: header
      name: x-session-token
      description: Pass the User ID / Session Token obtained from /auth/login
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          example: student@iiml.ac.in
        name:
          type: string
          example: Jane Doe
        role:
          type: string
          enum: [Booker, Approver, Admin]
          example: Booker

    AuthSuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: Authentication successful
        token:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        user:
          $ref: '#/components/schemas/User'